requirements.txt
pypot==3.1.3
flask==1.1.1
gcloud==0.18.3
sseclient==0.0.24
requests_toolbelt==0.9.1
python_jwt
pycrypto==2.6.1
prettytable
flask_cors
